## Continuing Eve tribute
Apparently, shooting for quantity over quality leads to better quality, so I've decided to complete a small project: the tribute to eve I started last month. Today I finished adding all of the images, videos and links to the page.

What remains:

* I need to decide if I should include the other links somewhere in the page, including other dev diaries, other Chris blogs, or Chris's other videos.
* The story to the page. If it's a "thanks for your hard work" or more of just an archieve, etc.
* Better CSS to make it sexy. Potentially, I'll want a better timeline-y format where the dates are given higher visual priority than they are now.

It'd be neat if I could get this out in the next few days, maybe on HN, Twitter (send a note to the Eve guys), and reddit on the approriate thread.