## Thoughts
### Eve Success

I was a little surprised to see my Eve project stay on the front page for basically a whole day! It got 150 points on HN and up to #6. I got over 10k people on the page, and around 500 on my website homepage. However, I didn't really get anything material out of this experiment.

It does pass my "do they spend more time on it then you do test", because I only spend 10-ish hours here, and even if the average time on the page was 1 minute, that's over 100 hours of time they spent on it.

### Offline morning

This morning I decided to work with my notebook for 2 hours offline, thinking about other projects to work on. I got a bit excited about a WoofJS refactor plan, but ultimately was most excited about an FRP Scratch tool that is similar to Facebook's Origami (which is based on Apple's Quartz Composer) in that you can see all the intermediate values flowing through, immediate updates, etc.

### Unproductive afternoon

After lunch, I was lazy and watched a few hours of TV. Then I got myself on my computer to look at my Cycle and Elm Flappy Birds for inspiration. I was quite fustrated with how awful coding in Cycle and Elm is compared to Facebook Origami, and also a bit inspired by that, too. However, I feel lazy today, so I'll have to pick up on this thread later. 

### Thoughts for next week

* Scratch FRP prototyping
* how do Facebook's Origami, CycleJS devtools, and statecharts somehow fit together? 
* WoofJS refactor plan (start with docs and a build tool)