## Dataflow reading
Inspired by my conversation with Brent Yorgey yesterday, I spent ~2 hours reading a 30-page survey paper on data flow programming langauges, with an emphasis on visual data flow langauages, such as LabView. It was fascinating!

A good next step would be to find a similar kind of survey paper but for functional reactive programming. While I could just start with Conal's Fran paper, I expect it'd be more fun and a better use of my time to get a more modern and broad perspective -- if it could include ReactJS and other JS libraries too, even better!

### Booleans are too generic - proofs are better

Brent reccomended [a great article](https://existentialtype.wordpress.com/2011/03/15/boolean-blindness/) in a similar spirit to my "booleans are too generic" idea. The main idea here is that truth shouldn't be a data type, but should be a proof, as it is in math. 

That's a bit different from my idea which proposes to create a lot of new, more specific data types to replace True and False, such as Equal and NotEqual, which would be pattern matched against, not branched at like an if-statement.