## Another vision-y essay
I started another essay today. This isn't a great sign that I keep creating new essays, instead of editing the old ones...

I'd like to find a way to release something, which means I'll have to cut down the scope as much as possible. The essay I worked on today does a good job of that. It's just trying to make the point that live modify-ability is a great thing. It's trying to sell the vision.

The other two essays: [visual](./essays/visual) and [casual](./essays/casual) and more about why FP and FRP is neccesary, and how we'll make it work. I'm stil not sure how to combine or seperate those essays.

Even worse, I've been collecting a large number of notes to add to these essays, and I don't know where they go...

* include referneces to http://glench.com/LegibleMathematics/ in my visualizable essay and also talk about how visualizations are key. they're already in our heads even if we don't know it.visual metaphors. maybe quote the monad guy. quotes about mathmaticians who see things. 
* also I can clarify that when trying to improve understandability of programming you can either improve the model or the interface, and how I'm working on both dimensions at the same time (paul chiusano leverage, glench is only on UI where I used to be, etc) 
* comprehensibility: model + interface...steps proj, Nile + Bret’s visualizations
* Haskell note from rattray: compiler knows more than you do
* declariative langauges stink but they allow for foundation for visual tools such as squarespace and looker