## Back at the FRP essay!
* TOC
{: toc }

The changes I'm comitting now were actually done this past Friday. It's not a ton of work but it wasn't pulling teeth - it was fun. And I'm excited enough about it that I actually made my girlfriend listen to me read it to her - she was a great sport. The langauge is a bit technical / high level. There's a lot I need to expound on / elaborate in there, and I imagine that some ambiguity in the essay are weak spots I need to explore more. 

Yesterday I had a lot of fun comparing and contrasting TodoMVC in Elm and Redux. I printed out the code for both on paper and wrote all over them. I have a lot of questions about the types of things in Redux that I'll have to investigate next.

One idea I had is that to make the comparison more apples-to-apples, I could write TodoMVC in Redux but with the Elm Architecture - this is possible but not the reverse. However, that is a bunch of work considering how slow my Reflex/ghcjs feedback loop is. Another possible task is implementing focus on input elements, localstorage, and url hash get/set in Reflex to make them truly equal - but that's not really the point so it may be more work than it's worth.

I only have two weeks before I have to submit to REBLS. I think I have enough for the paper in progress section, but I want to make it as best I can before that date. However, I also have work for Dark and First Round I should do this week - as well as taking some time off to spend with family that's around this month. It's a lot going on!

### Dark research

I spent 3 hours this morning compiling lists of various interesting progrmming environments. What I'm most excited about is my list of classics that I've never played with:

* Hypercard
* squeak / etoys / morphicjs / pharo / http://gtoolkit.org/
* Labview
* apl
* Self
* Forth
* Prolog
* Sketchpad
* Matlab / simulink
* Yahoo pipes


As well as a few I keep haering about:

* prograph
* paxmsp and pure data  vvvv
* JetBrains MPS 
* emacs org mode 
* TLA+ / Alloy
* Sonic Pi 
* R
* https://www.blender.org/
* Godot engine
* Qt
* Chalktalk
* http://mbeddr.com/
* http://openendedgroup.com/field/OverviewBanners2.html