## London, REBLS, & future plans
* TOC
{: toc }


### Moving to London!

Last week I got the Exceptional Tech Talent Visa which means I'm moving there on Oct 1. Just bought my plane ticket. Hope to be there for a year or two.

### Submitted to REBLS 2018!!

Last night I submitted my the [the submitted PDF version](./papers/comprehensible-frp/comprehensible-frp.pdf) of the paper to REBLS! It took ~10 hours to translate the [blog version](./papers/comprehensible-frp) to the PDF version! I had a lot of trouble with LaTeX, so I fiuged I should put up the [full LaTex code and associated files ](https://github.com/stevekrouse/futureofcoding.org/tree/master/papers/comprehensible-frp/LaTex) in case it'd be helpful for future independent researchers. Turns out they extended the deadline for another two weeks just now (probably because they didn't get enough submissions), but it's nice to have it behind me.

Looking at the cost of flights to Boston from London, it seems pretty stable, so I don't need to get it yet. I can also register for SPLASH up until Oct 1 for $100 off. I hear back if I get in on Sept 25th.

### Next two weeks (8/20/18-8/31/18)

Next week I agreed to write an essay about Dynamicland. Also, I'm visiting Boston for a few days to visit my programming research friends! I'd also like to do more work for First Round Capital next week. So I don't think I'll have much time here.

The following week will be more vacation focused with my parents, brothers, and grandparents. I may do a bit of work for First Round and Dark, but not likely work here.

### Future research directions

Here are a few ideas. I'm going to chat with Jonathan Edwards in Boston next week so maybe we can review this together.

#### 1. [Comprehensible FRP's Future work](./papers/comprehensible-frp#7-future-work)

I'm actually very excited about this section. Here's basically the plan. Take the Reflex TodoMVC diagram from this paper and tweak in in Figma a bunch. At the same time, work on getting this stuff to compile in JavaScript somehow. The goal here is very explicitly a language / visual environment. If it's good, I win. If it's bad, maybe I turn it into a paper.

#### 2. [Anti-IO Monad](http://futureofcoding.org/log#my-growing-anti-io-monad-obsession)

This is a neat line of inquiry but I don't have a very concrete next step here. It's very related to the ideas in this tweet storm:

<div id="tttt_1027765242217029632" data-option="1"><strong><a href="https://threadreaderapp.com/thread/1027765242217029632.html">Thread by @stevekrouse: "1/7 If there’s more than one way to encode an idea (in fully reduced form), you’re not being abstract/specific enough. Your encoding is loss […]"</a></strong></div><script async src="https://threadreaderapp.com/embed/1027765242217029632.js" charset="utf-8"></script>

And the lovely replies: 

<blockquote class="twitter-tweet" data-lang="en"><p lang="en" dir="ltr">Interesting.<br><br>2¢ — I think when you push for extreme reductions, throwing away ever more context.. you eventually run into the Berry Paradox (<a href="https://t.co/4ULmDa7DPt">https://t.co/4ULmDa7DPt</a>)</p>&mdash; Ivan Reese (@spiralganglion) <a href="https://twitter.com/spiralganglion/status/1027776456808521728?ref_src=twsrc%5Etfw">August 10, 2018</a></blockquote>
<script async src="https://platform.twitter.com/widgets.js" charset="utf-8"></script>

<blockquote class="twitter-tweet" data-lang="en"><p lang="en" dir="ltr">sounds like church encoding <a href="https://t.co/4Md0GECpXo">https://t.co/4Md0GECpXo</a></p>&mdash; Mariano Guerra (@warianoguerra) <a href="https://twitter.com/warianoguerra/status/1027956450528841728?ref_src=twsrc%5Etfw">August 10, 2018</a></blockquote>
<script async src="https://platform.twitter.com/widgets.js" charset="utf-8"></script>

#### 3. [Slice-and-Dice Data-ninja playground](http://futureofcoding.org/log#yesterday%E2%80%99s-slice-and-dice-data-ninja-playground)

This is very much jumping to an entirely different problem. It's intriguing, but not as likely.

#### 4. Liberal Software

From a motivation/vision perspective, this is important. I'd love to collaborate with r0ml on this.