## Notes 10/31/18-11/1/18
* TOC
{: toc }

### More FRP talk practice!

[Yesterday's practice](https://www.useloom.com/share/d52de5de266b4862afa2777176f652bb) was pretty solid! I'm not sure if I'll have time to review it and do a better one today...

### Stephen Diehl lunch

This morning I poked around on the internet for tutorials related to creating "toy Haskells" particularly interpreted and not compiled. This is related to wanting to create a FP language with a much *livelier* feel, aka [prototype 4](https://futureofcoding.org/log#prototype-4) and sometimes referred to as viv, potluck, stride, etc. I came across Stephen Diehl's many writings on these topics and saw on his Twitter that he went to the London Haskell meetup last night, so I messaged him and we had lunch a few hours later! I love the internet.

He claimed that in order to design a functional language you need to settle on three things:

1. AST
2. Core calculi
3. Evaluation semantics

The code is parsed to the AST, the AST is reduced to something that resembles the core calculi (you can see this in ghci with D dump core), and then the evaluation semantics determine how things are evaluated.

#### Core calculi

* untyped lambda calculus (Scheme)
* System-F (Haskell)
* Martin Lof (Haskell + dependent types)

#### Evaluation semantics

1. Call by value (eval arguments first)
2. Call by need (lazy, like Haskell)
3. Call by push value (he didn't explain)

#### Random notes

* Apparently recursive-do is possible in Stephan's basic interpreter in his book (2k lines of Haskell) because it's just built on top of mfix, which is mutual recursion, and bind...
* He strongly reccomended I read Ben Pierce's book, at least chapters 1-11. One of his coworkers forewarned that it's more of a reference manual than a book to read cover to cover.
* He also recommended Robert Harper's _Practical Foundations_, which is about adding things a la carte to semantics.

### Splash plans

Dinner Thursday night with some folks. Message me if you want to join!

### Interesting cloud data projects

I found [edsu.org](https://edsu.org/) via [this HN post](https://news.ycombinator.com/item?id=18316116). Seems cool. Similar to a lot of other projects floating around the ether these days. The comments are full of them. Just wanted to jot it down here.

### Todos 11/1/18

* continue research with [prototype 4](https://futureofcoding.org/log#prototype-4) (including learning about the stuff from Stephan Diehl above) and/or [multi-node FRP](https://github.com/stevekrouse/futureofcoding.org/issues/85) 
* Read all of [Kartik](http://akkartik.name/) & chat with Antranig and Philip & redo my /about page
* Organize FoC Thinking & FoC Research lists from Google Inbox --> github issues/project (including these todos)